refactor(emails): move the test module into a sibling file
emails.rs had grown to 4,629 lines with roughly 60 percent of it tests,
which buried the production code it was meant to sit beside.

The tests move to emails_tests.rs, wired with the path attribute this
crate already uses in user.rs, user_role.rs, coaching_session_topic.rs,
and gateway/ical.rs. emails.rs itself already extracted one test module
this way, so the split follows a pattern the file had started.

It also lets the test file be made read-only for frozen-test work without
locking the source next to it.

Pure move: the diff to emails.rs is the three wiring lines and the
deletion, no production code touched, and the domain suite reports the
same 302 tests before and after.
